Modifications for compatibility and uniformity oc complex types
This PR makes minor modifications to improve compatibility across the codebase.
The changes focus on correcting data type usage and updating error messages for consistency.

Names of elements of complex types made uniform across code base.
Whole code-base uses now {real,imag}.
Fixed data type mismatch in BSVD sorting function
Updated error messages to use consistent terminology
Removed unnecessary macro undefining

+++ src/base/flamec/util/lapack/FLA_Sort_bsvd_ext.c
@@ -208,7 +208,7 @@ FLA_Error FLA_Sort_bsvd_ext_f_opd( fla_dim_t m_s, double* s, fla_dim_t inc_s,
                                    fla_dim_t n_C, double* C, fla_dim_t rs_C, fla_dim_t cs_C )
 {
     fla_dim_t    i, ii, j, k;
-    float  p;
+    double  p;
     FLA_SORT_BSVD_EXT_BODY( FORWARD, bl1_dswapv );
     return FLA_SUCCESS;
 }
